refactor(test): split test_governance.py to reduce file size
Split tests/vibe3/roles/test_governance.py (578 lines, exceeds 500 limit)
into two focused test files:

- test_governance_context.py (354 lines): TestBuildSnapshotContext class
  - Tests for build_governance_snapshot_context() function
  - Includes new vibe-task filtering tests added in PR #949

- test_governance.py (268 lines): Remaining 7 test classes
  - TestBuildGovernanceRecipe
  - TestRenderGovernancePrompt
  - TestBuildGovernanceRequest
  - TestBuildExecutionName
  - TestGovernanceMaterials
  - TestRoundRobinMaterialSelection
  - TestGovernanceRoleDefinition

Rationale:
- Both files now below 500-line LOC limit
- TestBuildSnapshotContext focuses on context building (single responsibility)
- Other tests remain grouped as they share common governance workflow logic
- No functional changes, only file restructuring
- All 30 tests pass successfully

Related to PR #949 (issue #865 vibe-task filtering implementation)
